Moved updater
