+++ atcoder-problems-backend/src/lib.rs
@@ -5,3 +5,4 @@ pub mod s3;
 pub mod scraper;
 pub mod sql;
 pub mod error;
+pub mod updater;
