--- atcoder-problems-updater-lambda/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "atcoder-problems-updater-lambda"
-version = "0.1.0"
-authors = ["kenkoooo <kenkou.n@gmail.com>"]
-edition = "2018"
-
-[dependencies]
-regex = "1"
-atcoder-problems-sql-common = { path = "../atcoder-problems-sql-common" }
-lambda_runtime = "^0.1"
-diesel = { version = "1.4.2", features = ["postgres"] }
-openssl-probe = "0.1.2"
-openssl = "*"
-log = "^0.4"
-simple_logger = "^1"
