--- atcoder-problems-updater-lambda/src/lib.rs
@@ -1,2 +1,0 @@
-pub mod updater;
-pub mod delta_updater;
